% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{auprc.pvrandom}
\alias{auprc.pvrandom}
\title{Assess the significance of AUPRC against random guessing
#'}
\usage{
auprc.pvrandom(yr2, monotonized = TRUE, cycles = 10000)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{monotonized}{whether or not to monotonize the curves}

\item{cycles}{the sample size of the null distribution to use}
}
\value{
The emprical p-values of the AUPRC against random guessing
}
\description{
Assess the significance of AUPRC against random guessing
#'
}
\examples{
#generate fake data
N <- 10
M <- 8
truth <- c(rep(TRUE,N),rep(FALSE,M))
scores <- cbind(
  pred1=c(rnorm(N,1,0.2),rnorm(M,.9,0.1)),
  pred2=c(rnorm(N,1.1,0.2),rnorm(M,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#call pvrandom function
auprc.pvrandom(yrobj)
}
