% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{auprc}
\alias{auprc}
\title{Calculate area under precision recall curve (AUPRC)}
\usage{
auprc(yr2, monotonized = TRUE, balanced = FALSE)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{monotonized}{whether or not use a monotonized PRC curve}

\item{balanced}{whether or not to use prior-balancing}
}
\value{
a numerical vector with the AUPRC values for each predictor
}
\description{
Concept by Yingzhou Wu and Fritz Roth (Wu et al, unpublished)
}
\examples{
#generate fake data
truth <- c(rep(TRUE,10),rep(FALSE,8))
scores <- cbind(
  pred1=c(rnorm(10,1,0.2),rnorm(8,.9,0.1)),
  pred2=c(rnorm(10,1.1,0.2),rnorm(8,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#calculate AUPRC
auprc(yrobj)
#calculate non-monotonized AUPRC
auprc(yrobj,monotonized=FALSE)
#calculate balanced AUPRC
auprc(yrobj,balanced=TRUE)
}
