% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{auprc.signif}
\alias{auprc.signif}
\title{Assess the significance of AUPRC differences}
\usage{
auprc.signif(yr2, monotonized = TRUE, res = 0.001)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{monotonized}{whether or not to monotonize the curve}

\item{res}{the resolution at which to sample the probability function 
(defaults to 0.001)}
}
\value{
a list containing 4 elements: "auprc" (the empirical area under the 
precision recall curve), "ci" (the 95%confidence interval around the auprc), 
"llr" (the log likelihood ratio matrix, see details), and "pval" (the p-value 
of each auprc against each other)
}
\description{
The list returned by this functions contains four elements:
\describe{
\item{auprc} is simply the empirical area under the precision recall curve 
for each predictor.
\item{ci} is a matrix listing the lower and upper end of the 95% confidence 
interval for the AUPRC of each predictor.
\item{llr} is a matrix with columns and rows corresponding to each predictor.
It lists the log likelihood ratio of how much more (or less) likely the row-wise
predictor is to have a greater AUPRC than the column-wise predictor.
\item{pval} is a matrix with columns and rows corresponding to each predictor.
It lists the p-value of how likely it would be to observe the AUPRC of the row-wise
predictor under the distribution of the column-wise predictor.
}
}
