% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{auroc}
\alias{auroc}
\title{Calculate the area under the ROC curve}
\usage{
auroc(yr2)
}
\arguments{
\item{yr2}{the yogiroc2 object}
}
\value{
a numerical vector with the AUROC for each predictor
}
\description{
Calculate the area under the ROC curve
}
\examples{
#generate fake data
truth <- c(rep(TRUE,10),rep(FALSE,8))
scores <- cbind(
  pred1=c(rnorm(10,1,0.2),rnorm(8,.9,0.1)),
  pred2=c(rnorm(10,1.1,0.2),rnorm(8,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#calculate AUROC
auroc(yrobj)
}
