% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{calc.auc}
\alias{calc.auc}
\title{Helper function to calculate area under curve}
\usage{
calc.auc(xs, ys)
}
\arguments{
\item{xs}{the x values of the graph}

\item{ys}{the corresponding y values of the graph}
}
\value{
the area under the curve
}
\description{
Helper function to calculate area under curve
}
