% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{calculate_thresh_range}
\alias{calculate_thresh_range}
\title{Extract thresholds}
\usage{
calculate_thresh_range(yr2, x = 0.9, monotonized = TRUE, balanced = FALSE)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{x}{the precision cutoff (default 0.9)}

\item{monotonized}{whether or not to use monotonized PRC}

\item{balanced}{whether or not to use prior-balancing}
}
\value{
threshold ranges
}
\description{
Extract thresholds
}
