% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{draw.prc.CI}
\alias{draw.prc.CI}
\title{Draw Precision-Recall Curve (PRC) with confidence intervals}
\usage{
draw.prc.CI(
  yr2,
  col = seq_along(yr2),
  monotonized = TRUE,
  legend = "bottomleft",
  ...
)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{col}{vector of colors to use for the predictors}

\item{monotonized}{whether or not to monotonize the curve}

\item{legend}{the position of the legend, e.g. "bottomleft". NA disables legend}

\item{...}{additional graphical parameters (see \code{par})}
}
\value{
nothing. draws a plot
}
\description{
Draw Precision-Recall Curve (PRC) with confidence intervals
}
\examples{
#generate fake data
N <- 100
M <- 80
truth <- c(rep(TRUE,N),rep(FALSE,M))
scores <- cbind(
  pred1=c(rnorm(N,1,0.2),rnorm(M,.9,0.1)),
  pred2=c(rnorm(N,1.1,0.2),rnorm(M,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#draw PRC curve
draw.prc.CI(yrobj)
#draw non-monotonized PRC curve
draw.prc.CI(yrobj,monotonized=FALSE)
}
