% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{draw.prc}
\alias{draw.prc}
\title{Draw Precision-Recall Curve (PRC)}
\usage{
draw.prc(
  yr2,
  col = seq_along(yr2),
  lty = 1,
  monotonized = TRUE,
  balanced = FALSE,
  legend = "bottomleft",
  ...
)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{col}{vector of colors to use for the predictors}

\item{monotonized}{whether or not to monotonized the curve}

\item{balanced}{whether or not to use prior-balancing}

\item{legend}{the position of the legend, e.g. "bottomleft". NA disables legend}

\item{...}{additional graphical parameters (see \code{par})}
}
\value{
nothing. draws a plot
}
\description{
Balancing concept by Yingzhou Wu and Fritz Roth (Wu et al, unpublished)
}
\examples{
#generate fake data
truth <- c(rep(TRUE,10),rep(FALSE,8))
scores <- cbind(
  pred1=c(rnorm(10,1,0.2),rnorm(8,.9,0.1)),
  pred2=c(rnorm(10,1.1,0.2),rnorm(8,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#draw PRC curve
draw.prc(yrobj)
#draw non-monotonized PRC curve
draw.prc(yrobj,monotonized=FALSE)
#draw balanced PRC curve
draw.prc(yrobj,balanced=TRUE)
}
