% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{draw.roc}
\alias{draw.roc}
\title{Draw a ROC curve}
\usage{
draw.roc(yr2, col = seq_along(yr2), lty = 1, legend = "bottomright", ...)
}
\arguments{
\item{yr2}{an underlying yogiroc2 object}

\item{col}{the colors to use for the predictors}

\item{legend}{the positioning of the legend (e.g. "bottomright). NA to disable legend.}

\item{...}{additional graphical parameters (see \code{par})}
}
\value{
nothing, draws a plot
}
\description{
Draw a ROC curve
}
\examples{
#generate fake data
truth <- c(rep(TRUE,10),rep(FALSE,8))
scores <- cbind(
  pred1=c(rnorm(10,1,0.2),rnorm(8,.9,0.1)),
  pred2=c(rnorm(10,1.1,0.2),rnorm(8,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#draw PRC curve
draw.roc(yrobj)
}
