% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{inferPRCCI}
\alias{inferPRCCI}
\title{Use the output of samplePRCs to infer confidence intervals for a PRC curve}
\usage{
inferPRCCI(randomPaths, nbins = 50)
}
\arguments{
\item{randomPaths}{the output of samplePRCs}

\item{nbins}{the number of bins along the recall axis to use}
}
\value{
a table listing the confidence interval at each recall bin
}
\description{
Use the output of samplePRCs to infer confidence intervals for a PRC curve
}
