% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{monotonize}
\alias{monotonize}
\title{Helper function to monotonize precision}
\usage{
monotonize(xs)
}
\arguments{
\item{xs}{numerical input vector, representing precision ordered according to increasing t}
}
\value{
the monotonized equivalent vector
}
\description{
Helper function to monotonize precision
}
