% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiroc.R
\name{new.yogiroc}
\alias{new.yogiroc}
\title{new yogiroc}
\usage{
new.yogiroc(truth, scores)
}
\arguments{
\item{truth}{a boolean vector for the ground truth classification}

\item{scores}{the scores to be evaluated agains the ground truth}
}
\value{
A new yogiroc object
}
\description{
Warning: This class is now deprecated. Please use \code{yr2} instead.
This creates a new yogiroc object used to draw ROC and PRC curves and calculate the respective AUC.
}
\details{
The object exposes the following methods:
\begin{description}
\item{draw.roc(col,main,lwd,add)} Draws a ROC curve. Parameter "col" controls the line color, 
"main" sets the plot title, "lwd" sets the line width, "add" allows the plot to be added as a line
to an existing plot.
\item{draw.prc(col,main,lwd,add)} Draws a PRC curve, Parameters as above.
\item{auroc()} Returns the Area under the ROC curve.
\item{auprc()} Returns the Area under the PRC curve.
\end{description}
}
