% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{prcCI}
\alias{prcCI}
\title{confidence interval for precision or recall}
\usage{
prcCI(i, n, p = c(0.025, 0.975), res = 0.001)
}
\arguments{
\item{i}{the number of correct calls}

\item{n}{the number of total calls}

\item{p}{the confidence interval probability range e.g. 0.025;0.975}

\item{res}{the resolution at which to sample the call rates}
}
\value{
the confidence interval (numerical vector)
}
\description{
confidence interval for precision or recall
}
