% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{print.yr2}
\alias{print.yr2}
\title{print method for yogiroc2 objects}
\usage{
\method{print}{yr2}(yr2)
}
\arguments{
\item{yr2}{the object}
}
\value{
nothing, just prints a description
}
\description{
print method for yogiroc2 objects
}
