% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{recall.at.prec}
\alias{recall.at.prec}
\title{Calculate maximum recall at given minimum precision}
\usage{
recall.at.prec(yr2, x = 0.9, monotonized = TRUE, balanced = FALSE)
}
\arguments{
\item{yr2}{the yogiroc2 object}

\item{x}{the precision cutoff (default 0.9)}

\item{monotonized}{whether or not to use monotonized PRC}

\item{balanced}{whether or not to use prior-balancing}
}
\value{

}
\description{
Calculate maximum recall at given minimum precision
}
\examples{
#generate fake data
truth <- c(rep(TRUE,10),rep(FALSE,8))
scores <- cbind(
  pred1=c(rnorm(10,1,0.2),rnorm(8,.9,0.1)),
  pred2=c(rnorm(10,1.1,0.2),rnorm(8,.9,0.2))
)
#create yogiroc2 object
yrobj <- yr2(truth,scores)
#calculate R90P
recall.at.prec(yrobj)
#calculate non-monotonized R90P
recall.at.prec(yrobj,monotonized=FALSE)
#calculate balanced R90P
recall.at.prec(yrobj,balanced=TRUE)
}
