% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{samplePRCs}
\alias{samplePRCs}
\title{Sample a distribution of PRC paths based, based on likelihood dictated by data}
\usage{
samplePRCs(data, N = 1000, monotonized = TRUE, sr = sampleRates)
}
\arguments{
\item{data}{a data table from a yr2 object}

\item{N}{the number of samples}

\item{monotonized}{whether to use monotonization}
}
\value{
a list of tables containin N samples of precision and recall each. 
   Each table corresponds to one row in the 'data' input
}
\description{
Sample a distribution of PRC paths based, based on likelihood dictated by data
}
