% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2.R
\name{sampleRatesQD}
\alias{sampleRatesQD}
\title{Quick and dirty sampling of rate parameters of a binomial distribution}
\usage{
sampleRatesQD(i, n, N = 1000, minQ = 0, maxQ = 1)
}
\arguments{
\item{i}{numerator (number of successful events)}

\item{n}{denominator (total number of events)}

\item{N}{desired number of samples to generate}

\item{minQ}{a minimum constraint for the samples 
Rates can't be smaller than this. This is useful to include prior information.}

\item{maxQ}{a maximum constarint for the samples}
}
\value{
a vector of sampled rates
}
\description{
This function is used internally by samplePRCs() (below)
}
